+++ platform/viewer/src/App.js
@@ -31,7 +31,8 @@ import {
 import i18n from '@ohif/i18n';
 
 // TODO: This should not be here
-import './config';
+//import './config';
+import { setConfiguration } from './config';
 
 /** Utils */
 import {
@@ -50,7 +51,7 @@ import { getActiveContexts } from './store/layout/selectors.js';
 import store from './store';
 
 /** Contexts */
-import WhiteLabellingContext from './context/WhiteLabellingContext';
+import WhiteLabelingContext from './context/WhiteLabelingContext';
 import UserManagerContext from './context/UserManagerContext';
 import AppContext from './context/AppContext';
 
@@ -77,7 +78,9 @@ class App extends Component {
       PropTypes.shape({
         routerBasename: PropTypes.string.isRequired,
         oidc: PropTypes.array,
-        whiteLabelling: PropTypes.object,
+        whiteLabeling: PropTypes.shape({
+          createLogoComponentFn: PropTypes.func,
+        }),
         extensions: PropTypes.array,
       }),
     ]).isRequired,
@@ -87,7 +90,6 @@ class App extends Component {
   static defaultProps = {
     config: {
       showStudyList: true,
-      whiteLabelling: {},
       oidc: [],
       extensions: [],
     },
@@ -107,7 +109,6 @@ class App extends Component {
       cornerstoneExtensionConfig: {},
       extensions: [],
       routerBasename: '/',
-      whiteLabelling: {},
     };
 
     this._appConfig = {
@@ -122,6 +123,8 @@ class App extends Component {
       extensions,
       oidc,
     } = this._appConfig;
+
+    setConfiguration(this._appConfig);
 
     this.initUserManager(oidc);
     _initServices([
@@ -146,7 +149,7 @@ class App extends Component {
   }
 
   render() {
-    const { whiteLabelling, routerBasename } = this._appConfig;
+    const { whiteLabeling, routerBasename } = this._appConfig;
     const {
       UINotificationService,
       UIDialogService,
@@ -162,7 +165,7 @@ class App extends Component {
               <OidcProvider store={store} userManager={this._userManager}>
                 <UserManagerContext.Provider value={this._userManager}>
                   <Router basename={routerBasename}>
-                    <WhiteLabellingContext.Provider value={whiteLabelling}>
+                    <WhiteLabelingContext.Provider value={whiteLabeling}>
                       <SnackbarProvider service={UINotificationService}>
                         <DialogProvider service={UIDialogService}>
                           <ModalProvider
@@ -175,7 +178,7 @@ class App extends Component {
                           </ModalProvider>
                         </DialogProvider>
                       </SnackbarProvider>
-                    </WhiteLabellingContext.Provider>
+                    </WhiteLabelingContext.Provider>
                   </Router>
                 </UserManagerContext.Provider>
               </OidcProvider>
@@ -190,7 +193,7 @@ class App extends Component {
         <Provider store={store}>
           <I18nextProvider i18n={i18n}>
             <Router basename={routerBasename}>
-              <WhiteLabellingContext.Provider value={whiteLabelling}>
+              <WhiteLabelingContext.Provider value={whiteLabeling}>
                 <SnackbarProvider service={UINotificationService}>
                   <DialogProvider service={UIDialogService}>
                     <ModalProvider modal={OHIFModal} service={UIModalService}>
@@ -198,7 +201,7 @@ class App extends Component {
                     </ModalProvider>
                   </DialogProvider>
                 </SnackbarProvider>
-              </WhiteLabellingContext.Provider>
+              </WhiteLabelingContext.Provider>
             </Router>
           </I18nextProvider>
         </Provider>